src/verisk_pipeline/requirements.txt
# Core dependencies
polars>=0.20.0
deltalake>=0.15.0
pyyaml>=6.0
python-dotenv>=1.0.0
tzdata>=2024.1
requests>=2.28.0
pyarrow>=14.0.0
psutil>=5.9.0


# Azure SDK
azure-kusto-data>=4.2.0
azure-identity>=1.15.0
azure-storage-file-datalake>=12.14.0

# Async HTTP
aiohttp>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0